noop-move waves to separate module
